+++ src/day9.c
@@ -8,13 +8,13 @@
 #define PREVIOUS_SIZE 25
 
 int u64_compare(const void *a, const void *b, void *udata) {
-    const unsigned long long *elm1 = a;
-    const unsigned long long *elm2 = b;
+    const u64 *elm1 = a;
+    const u64 *elm2 = b;
     return *elm1 - *elm2;
 }
 
 uint64_t u64_hash(const void *item, uint64_t seed0, uint64_t seed1) {
-    const unsigned long long *elm = item;
+    const u64 *elm = item;
     return *elm; // identity is a hash function :p
 }
 
@@ -30,13 +30,13 @@ void day9() {
             input_size++;
         }
     }
-    unsigned long long *input = MEM_alloc(input_size * sizeof (unsigned long long));
+    u64 *input = MEM_alloc(input_size * sizeof (u64));
     const u8 *cursor = DAY9_INPUT;
     for (u16 idx = 0; idx < input_size; idx++) {
         while (!isdigit(*cursor)) {
             cursor++;
         }
-        unsigned long long parsed_number = skip_atoi_u64(&cursor);
+        u64 parsed_number = skip_atoi_u64(&cursor);
         input[idx] = parsed_number;
     }
     sprintf(buf, "Parsed day 9 input in %lu ms", getTimer(0, FALSE) / SUBTICKPERMILLISECOND);
@@ -45,20 +45,20 @@ void day9() {
     // solve
     startTimer(0);
     drawText("Solving part 1...", 1, line++);
-    unsigned long long previous[PREVIOUS_SIZE];
+    u64 previous[PREVIOUS_SIZE];
     u16 previous_offset = 0;
     for (u16 i = 0; i < PREVIOUS_SIZE; i++) {
         previous[i] = input[i];
     }
     
-    unsigned long long invalid_number = 0;
-    struct hashmap *values_map = hashmap_new(sizeof (unsigned long long), 128, 0, 0, u64_hash, u64_compare, NULL);
+    u64 invalid_number = 0;
+    struct hashmap *values_map = hashmap_new(sizeof (u64), 128, 0, 0, u64_hash, u64_compare, NULL);
     for (u16 x = 0; x < PREVIOUS_SIZE; x++) {
         hashmap_set(values_map, &previous[x]);
     }
     for (u16 i = PREVIOUS_SIZE; i < input_size; i++) {
         for (u16 x = 0; x < PREVIOUS_SIZE; x++) {
-            unsigned long long value_to_find = input[i] - previous[x];
+            u64 value_to_find = input[i] - previous[x];
             if (value_to_find != previous[x] && hashmap_get(values_map, &value_to_find)) {
                 hashmap_delete(values_map, &previous[previous_offset]);
                 hashmap_set(values_map, &input[i]);
@@ -84,7 +84,7 @@ valid:
     drawText("Solving part 2...", 1, line++);
     u16 index_start = 0;
     u16 index_end = 1;
-    unsigned long long sum = input[index_start] + input[index_end];
+    u64 sum = input[index_start] + input[index_end];
     while (sum != invalid_number) {
         while (sum < invalid_number) {
             sum += input[++index_end];
@@ -93,8 +93,8 @@ valid:
             sum -= input[index_start++];
         }
     }
-    unsigned long long min = 0xffffffffffffffff;
-    unsigned long long max = 0;
+    u64 min = 0xffffffffffffffff;
+    u64 max = 0;
     for (u16 i = index_start; i <= index_end; i++) {
         if (input[i] > max) {
             max = input[i];
